Renamed push away auto function and moved out of auto commands
