+++ src/main/java/frc/robot/commands/AutoCommands/AutoPush.java
@@ -6,6 +6,7 @@ package frc.robot.commands.AutoCommands;
 
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import frc.robot.commands.ActuatorDown;
+import frc.robot.commands.PushAway;
 import frc.robot.commands.MoveCommand;
 import frc.robot.subsystems.DrivetrainSubsystem;
 import frc.robot.subsystems.IntakeSubsystem;
@@ -21,7 +22,7 @@ public class AutoPush extends SequentialCommandGroup {
     addCommands(
       new ActuatorDown(intake),
       parallel(new MoveCommand(drive, -7, false),
-      new AutoPushAway(intake))
+      new PushAway(intake))
     );
   }
 }
